+++ recognize_digits/.gitignore
@@ -2,3 +2,5 @@ data/raw_data
 data/train.list
 data/test.list
 *.log
+*.pyc
+plot.png
