+++ recognize_digits/image/max_pooling.png
