+++ recognize_digits/mnist_provider.py
@@ -13,36 +13,25 @@
 # limitations under the License.
 
 from paddle.trainer.PyDataProvider2 import *
-import numpy
+import numpy as np
+import struct
 
 
 # Define a py data provider
 @provider(
     input_types={'pixel': dense_vector(28 * 28),
-                 'label': integer_value(10)},
-    cache=CacheType.CACHE_PASS_IN_MEM)
+                 'label': integer_value(10)})
 def process(settings, filename):  # settings is not used currently.
-    imgf = filename + "-images-idx3-ubyte"
-    labelf = filename + "-labels-idx1-ubyte"
-    f = open(imgf, "rb")
-    l = open(labelf, "rb")
+    with open(filename + "-images-idx3-ubyte", "rb") as f:  # open picture file
+        magic, n, rows, cols = struct.unpack(">IIII", f.read(16))
+        images = np.fromfile(
+            f, 'ubyte',
+            count=n * rows * cols).reshape(n, rows, cols).astype('float32')
+        images = images / 255.0 * 2.0 - 1.0  # normalized to [-1,1]
 
-    f.read(16)
-    l.read(8)
-
-    # Define number of samples for train/test
-    if "train" in filename:
-        n = 60000
-    else:
-        n = 10000
-
-    images = numpy.fromfile(
-        f, 'ubyte', count=n * 28 * 28).reshape((n, 28 * 28)).astype('float32')
-    images = images / 255.0 * 2.0 - 1.0
-    labels = numpy.fromfile(l, 'ubyte', count=n).astype("int")
+    with open(filename + "-labels-idx1-ubyte", "rb") as l:  # open label file
+        magic, n = struct.unpack(">II", l.read(8))
+        labels = np.fromfile(l, 'ubyte', count=n).astype("int")
 
     for i in xrange(n):
         yield {"pixel": images[i, :], 'label': labels[i]}
-
-    f.close()
-    l.close()
